+++ changelog.md
@@ -4,6 +4,7 @@
 * Experience
   * Experience from blocks config has been moved to BlockData
   * Fully Grown Crops now give 1~2 experience
+  * Amethyst clusters now give 1~2 experience
   * Harvesting Honey or Honeycombs from smoked Hives now awards 5 to 10 experience 
 * Reduced health on respawn